pyproject.toml
[tool.poetry]
name = "autofolio"
version = "0.1.0"
description = "From Code to Career - 증거 기반 개발자 이력서"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
requests = ">=2.28.0"
beautifulsoup4 = ">=4.11.0"
playwright = ">=1.40.0"
openai = ">=1.0.0"
python-dotenv = "^1.2.1"
langgraph = ">=0.2.0"
aiosqlite = "^0.22.1"
chromadb = "^1.5.5"
langchain-openai = "^1.1.11"
fastapi = "^0.135.1"

[tool.poetry.group.dev.dependencies]
# pytest = "^7.0.0"
# black = "^23.0.0"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
ruff = "^0.15.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
